#56 Rename helper.
