+++ lib/file-utils.js
