+++ lib/get-metadata-page.js
@@ -3,7 +3,7 @@
 const fs = require('fs');
 const path = require('path');
 const ejs = require('ejs');
-const getFilesContent = require('./helper').getFilesContent;
+const getFilesContent = require('./file-utils').getFilesContent;
 
 /**
  * Returns the html of the rendered metadata page.
