+++ lib/get-playground-page.js
@@ -2,7 +2,7 @@
 
 const ejs = require('ejs');
 const _ = require('lodash');
-const getFilesContent = require('./helper').getFilesContent;
+const getFilesContent = require('./file-utils').getFilesContent;
 
 /**
  * Returns the html of the rendered metadata page.
